优化编译警告

+++ core/plugins/src/lcodec/mysql.h
@@ -255,7 +255,7 @@ namespace lcodec {
             size_t column_count = decode_length_encoded_number();
             // field metadata
             mysql_columns columns;
-            for (int i = 0; i < column_count; ++i) {
+            for (size_t i = 0; i < column_count; ++i) {
                 recv_packet();
                 field_decode(columns);
             }
